fancy formatting math
