+++ src/selectMenus/sampleSelectMenu.ts
@@ -6,8 +6,8 @@ class ExampleSelectMenu extends SelectMenuHandler {
 		super("exampleSelectMenu");
 	}
 
-	async execute(interaction: SelectMenuInteraction, args: string[]): Promise<void> {
-		await interaction.reply(`You selected the following options: ${interaction.values}`);
+	async execute(interaction: SelectMenuInteraction, subcommand: string[]): Promise<void> {
+		await interaction.reply(`You used ${subcommand} and selected the following options: ${interaction.values}`);
 	}
 }
